+++ test/Main.hs
@@ -2,14 +2,12 @@
 {-# language ScopedTypeVariables #-}
 {-# language TypeApplications #-}
 
-import Data.Builder (singleton,run)
-import Control.Monad.ST (runST)
+import Data.Builder (singleton,doubleton,tripleton,run)
 import Test.Tasty (defaultMain,testGroup,TestTree)
 import Test.Tasty.HUnit ((@=?))
 import Data.Semigroup (stimes)
 import qualified Data.List as L
 import qualified Data.Foldable as F
-import qualified GHC.Exts as Exts
 import qualified Test.Tasty.HUnit as THU
 
 main :: IO ()
@@ -37,4 +35,17 @@ tests = testGroup "Tests"
     )
   , THU.testCase "C" $ (L.replicate 500 'X') @=?
     (F.toList $ run (stimes (500 :: Int) (singleton 'X')))
+  , THU.testCase "D" $ "ACDCDCDCDCDCDCDCDCDX" @=?
+    ( F.toList $ run
+      (  singleton 'A'
+      <> stimes (9 :: Int) (doubleton 'C' 'D')
+      <> singleton 'X'
+      )
+    )
+  , THU.testCase "E" $ "ABCABCABCABCABCABCABCX" @=?
+    ( F.toList $ run
+      (  stimes (7 :: Int) (tripleton 'A' 'B' 'C')
+      <> singleton 'X'
+      )
+    )
   ]