+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1.0 -- 2020-??-??
 
 * Add `new1`.
+* Add `Data.Builder.(doubleton|tripleton)`.
 
 ## 0.1.0.0 -- 2019-09-12